use correct type to fix a warning :P
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@9199 d7cf8633-e32d-0410-b094-e92efae38249

+++ cl_main.c
@@ -1231,7 +1231,7 @@ void CL_UpdateViewModel(void)
 	ent->state_previous = ent->state_current;
 	ent->state_current = defaultstate;
 	ent->state_current.time = cl.time;
-	ent->state_current.number = (unsigned int)-1;
+	ent->state_current.number = (unsigned short)-1;
 	ent->state_current.active = true;
 	ent->state_current.modelindex = cl.stats[STAT_WEAPON];
 	ent->state_current.frame = cl.stats[STAT_WEAPONFRAME];
